Simplify frontend builds and split coordinator
