+++ create_mountaineer_app/create_mountaineer_app/templates/project/[project_name]/app.py
@@ -1,5 +1,4 @@
 from mountaineer.app import AppController
-from mountaineer.client_compiler.postcss import PostCSSBundler
 from mountaineer.render import LinkAttribute, Metadata
 
 {% if create_stub_files %}
@@ -9,15 +8,10 @@ from {{project_name}}.controllers.home import HomeController
 from {{project_name}}.config import AppConfig
 
 controller = AppController(
-    config=AppConfig(), # type: ignore
-    {% if use_tailwind %}
+    config=AppConfig(),  # type: ignore
     global_metadata=Metadata(
         links=[LinkAttribute(rel="stylesheet", href="/static/app_main.css")]
     ),
-    custom_builders=[
-        PostCSSBundler(),
-    ],
-    {% endif %}
 )
 
 {% if create_stub_files %}
