+++ create_mountaineer_app/create_mountaineer_app/templates/project/[project_name]/views/package.json
@@ -13,12 +13,12 @@
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
-    "@types/react": "^19.0.10",
+    "@types/react": "^19.0.10"
     {% if use_tailwind %}
+    ,
     "tailwindcss": "^4.1.6",
     "@tailwindcss/postcss": "^4.1.6",
+    "postcss": "^8.4.33"
     {% endif %}
-    "postcss": "^8.4.33",
-    "postcss-cli": "^11.0.0"
   }
 }
