--- docs/api/build_plugins/base/page.mdx
@@ -1,5 +1,0 @@
-# Build Plugin Base
-
-Use the `ClientBuilderBase` to implement a new stage or custom handling during the frontend build lifecycle. Note that this API is still under development and might change between minor Mountaineer versions.
-
-::: mountaineer.client_compiler.base.ClientBuilderBase
