--- docs/api/build_plugins/postcss/page.mdx
@@ -1,3 +1,0 @@
-# PostCSS
-
-::: mountaineer.client_compiler.postcss.PostCSSBundler
