+++ docs/api/controllers/page.mdx
@@ -74,4 +74,4 @@ Every page on Mountaineer is server-side rendered by default. Most SSR logic is 
 
 ## Source Maps
 
-::: mountaineer.client_compiler.source_maps.SourceMapParser
+::: mountaineer.source_maps.SourceMapParser
