+++ docs/guides/deploy/page.mdx
@@ -10,10 +10,7 @@ First, add the following to your `.dockerignore` file. This will prevent Docker 
 
 ```
 **/node_modules
-**/_server
-**/_ssr
-**/_static
-**/_metadata
+**/.mountaineer
 ```
 
 ### Image 1: Frontend Dependencies
